# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocSholder API - Plataforma de Documentação Cirúrgica de Ombro e Cotovelo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients for the current doctor
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientWithSurgeries"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientWithSurgeries"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        nome:
          type: string
        dataNascimento:
          type: ["string", "null"]
        sexo:
          type: ["string", "null"]
        telefone:
          type: ["string", "null"]
        lado:
          type: ["string", "null"]
        nivelAtividade:
          type: ["string", "null"]
        beightonScore:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, doctorId, nome, createdAt]
    PatientWithSurgeries:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        nome:
          type: string
        dataNascimento:
          type: ["string", "null"]
        sexo:
          type: ["string", "null"]
        telefone:
          type: ["string", "null"]
        lado:
          type: ["string", "null"]
        nivelAtividade:
          type: ["string", "null"]
        beightonScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        surgeries:
          type: array
          items:
            $ref: "#/components/schemas/Surgery"
      required: [id, doctorId, nome, createdAt, surgeries]
    CreatePatientBody:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
        email:
          type: string
        sexo:
          type: string
        telefone:
          type: string
        lado:
          type: string
        nivelAtividade:
          type: string
        beightonScore:
          type: integer
        planoSaude:
          type: string
        indicadoPor:
          type: string
        endereco:
          type: string
        cidade:
          type: string
        estado:
          type: string
        cep:
          type: string
      required: [nome]
    UpdatePatientBody:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
        email:
          type: string
        sexo:
          type: string
        telefone:
          type: string
        lado:
          type: string
        nivelAtividade:
          type: string
        beightonScore:
          type: integer
        anamnese:
          type: string
          nullable: true
        laudos:
          type: string
          nullable: true
        planoSaude:
          type: string
        indicadoPor:
          type: string
        endereco:
          type: string
        cidade:
          type: string
        estado:
          type: string
        cep:
          type: string
    Surgery:
      type: object
      properties:
        regiao:
          type: ["string", "null"]
          enum: ["shoulder", "elbow", null]
        dadosClinicos:
          type: ["object", "null"]
          additionalProperties: true
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        dataCirurgia:
          type: ["string", "null"]
        hospital:
          type: ["string", "null"]
        tipoCaso:
          type: ["string", "null"]
        tiposProcedimento:
          type: array
          items:
            type: string
        procedimentoRealizado:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, patientId, doctorId, tiposProcedimento, createdAt]
...